+++ src/app/utils/utils.ts
@@ -33,9 +33,9 @@ export function isWeb(): boolean {
 }
 
 export function nowAtEvent(timeZone: string): Date {
-  // if (environment.simulatedTime) {
-  //   return clone(environment.simulatedTime);
-  // }
+  if (environment.simulatedTime) {
+    return clone(environment.simulatedTime);
+  }
   let str = new Date().toLocaleString('en-US', { timeZone });
   return new Date(str);
 }