+++ src/environments/environment.snrg.ts
@@ -2,11 +2,13 @@ import { Environment } from "./environment.model";
 
 export const environment: Environment = {
   production: true,
-  simulatedTime: undefined,
-  gps: undefined,
-  overrideLocations: false,
-  latitudeOffset: 0.81975,
-  longitudeOffset: -1.369113,
+  simulatedTime: new Date('8/27/2025, 7:07:09 PM'),
+  gps: { lng: -113.3722705, lat: 41.5800984 },
+  latitudeOffset: undefined,
+  longitudeOffset: undefined,
+  overrideLocations: true,
+  // latitudeOffset: 0.81975,
+  // longitudeOffset: -1.369113,
 };
 //36.994348, -116.717122 snrg
 //36.174598, -115.348009 home