+++ changelog.md
@@ -1,5 +1,9 @@
 ## Changes
 
+## 2.99
+- Fix for calendar export when offline
+- Feature to clear favorites
+
 ## 2.98
 - Show friends on the map
 - Option to only show non recurring events